moved api_test below spec and added _functions
